feat(module new): use the first module-dir by default if available
Or default to "." if not configured.
